+++ src/hooks/useDeleteAccount.ts
@@ -1,5 +1,3 @@
-'use client';
-
 import { useRouter } from 'next/navigation';
 import { useAuthStore } from '@/store/useAuthStore';
 import { userAPI } from '@api/functions/userAPI';
